pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "research-collector"
version = "0.1.0"
description = "前沿文献采集与知识架构系统"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "rapidfuzz>=3.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]
semantic = ["sentence-transformers>=2.7"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
